update readme with release and rename run_executable to sim_manager
